Add a service CSRF protection
